loader/requirements.txt
google-cloud-storage==2.16.0
google-cloud-bigquery==3.18.0
functions-framework==3.5.0
